+++ CHANGELOG.md
@@ -88,3 +88,8 @@ Initial release of project supporting ISIR Layout specification released today, 
 
 - Added the updated [system-generated ISIR file](https://fsapartners.ed.gov/knowledge-center/library/handbooks-manuals-or-guides/2023-05-31/2024-25-fafsa-specifications-guide-march-2024-update) to the [test-isir-files folder](/test-isir-files). Changes to the file are outlined in the March 8th updates on the [2024–25 FAFSA Updates page](https://fsapartners.ed.gov/knowledge-center/topics/fafsa-simplification-information/2024-25-fafsa-updates#pid_1399062) on the FSA Knowledge Center.
 
+
+## 2024-04-03
+
+### Added Tools
+- Added [`fafsa-uuid-search.html`](./fafsa-uuid/fafsa-uuid-search.html) to allow searching a CSV file of up to 10 million UUIDs.